+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 - Ravamp emulator networking to assign ports early and explictly listen on IP addresses (#5083).
 - Emulator UI and hub now listen on both IPv4 and IPv6 address by default (if possible) (#5088).
 - Fix Firestore emulator excessive logs about discovery endpoint not found (#5088).
+- Adds additional Hosting site information to `hosting:sites:list`.